moved form into create journal form
