+++ weight-lifting-journal-front-end/src/components/Forms/CreateJournal.js
@@ -1,9 +1,50 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-const CreateJournal = () => {
+
+const CreateJournal = ({addNewJournal, today}) => {
+
+    // State for form
+    const [journal, setJournal] = useState({
+        id: '',
+        userId: '',
+        date: today,
+        region: ''
+    })
+    // Handle input change
+    const handleChange = event => {
+        setJournal({
+            ...journal,
+            [event.target.name]: event.target.value
+        })
+        console.log(journal)
+    }
+    // Handle form submit
+    const handleSubmit = event => {
+        event.preventDefault();
+        addNewJournal(journal)
+        setJournal({
+            id: '',
+            userId: '',
+            date: today,
+            region: ''
+        })
+    }
+    
 
     return (
-        <div>Create Journal</div>
+        <div>
+            <h2>Add Journal Entry</h2>
+            <form onSubmit={handleSubmit}>
+                <label>Name:</label>
+                <input
+                onChange={handleChange}
+                name="region"
+                type="text"
+                value={journal.region}
+                />
+                <button type="submit">Submit</button>
+            </form>
+        </div>
     )
 }
